wip(index): Migrating to Slim3
