--- include/config.php
@@ -1,10 +1,0 @@
-<?php
-/**
- * Database configuration
- */
-define('DB_USERNAME', 'root');
-define('DB_PASSWORD', 'root');
-define('DB_HOST', 'localhost');
-define('DB_NAME', 'chart_db');
-
-?>
