--- index.php
@@ -1,11 +1,0 @@
-<?php
-error_reporting(-1);
-ini_set('display_errors', 'On');
-?>
-<html>
-    <head>
-    </head>
-    <body>
-        Hola soy un API
-    </body>
-</html>
