--- v1/.htaccess
@@ -1,3 +1,0 @@
-RewriteEngine On 
-RewriteCond %{REQUEST_FILENAME} !-f 
-RewriteRule ^(.*)$ %{ENV:BASE}index.php [QSA,L]
